requirements.txt
flask
python-dotenv
gunicorn==20.0.4
flask-sqlalchemy
flask-mashmallow